decoder support verify checksum
