+++ pkg/sink/codec/avro/decoder.go
@@ -18,8 +18,6 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
-	"hash/crc32"
-	"math"
 	"strconv"
 	"strings"
 	"time"
@@ -153,7 +151,7 @@ func (d *decoder) NextRowChangedEvent() (*model.RowChangedEvent, error) {
 	}
 
 	if found {
-		if err := d.verifyChecksum(event.Columns, expectedChecksum); err != nil {
+		if err := common.VerifyChecksum(event.Columns, expectedChecksum); err != nil {
 			return nil, errors.Trace(err)
 		}
 	}
@@ -471,134 +469,4 @@ func (d *decoder) decodeValue(ctx context.Context) (map[string]interface{}, map[
 	data := d.value
 	d.value = nil
 	return decodeRawBytes(ctx, d.schemaM, data, d.topic)
-}
-
-// calculate the checksum value, and compare it with the expected one, return error if not identical.
-func (d *decoder) verifyChecksum(columns []*model.Column, expected uint64) error {
-	checksum, err := calculateChecksum(columns)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if checksum != expected {
-		log.Error("checksum mismatch",
-			zap.Uint64("expected", expected),
-			zap.Uint64("actual", checksum))
-		return errors.New("checksum mismatch")
-	}
-
-	return nil
-}
-
-// calculate the checksum, caller should make sure all columns is ordered by the column's id.
-// by follow: https://github.com/pingcap/tidb/blob/e3417913f58cdd5a136259b902bf177eaf3aa637/util/rowcodec/common.go#L294
-func calculateChecksum(columns []*model.Column) (uint64, error) {
-	var (
-		checksum uint32
-		err      error
-	)
-	buf := make([]byte, 0)
-	for _, col := range columns {
-		if len(buf) > 0 {
-			buf = buf[:0]
-		}
-		buf, err = buildChecksumBytes(buf, col.Value, col.Type)
-		if err != nil {
-			return 0, errors.Trace(err)
-		}
-		checksum = crc32.Update(checksum, crc32.IEEETable, buf)
-	}
-	return uint64(checksum), nil
-}
-
-// buildChecksumBytes append value the buf, mysqlType is used to convert value interface to concrete type.
-// by follow: https://github.com/pingcap/tidb/blob/e3417913f58cdd5a136259b902bf177eaf3aa637/util/rowcodec/common.go#L308
-func buildChecksumBytes(buf []byte, value interface{}, mysqlType byte) ([]byte, error) {
-	if value == nil {
-		return buf, nil
-	}
-
-	switch mysqlType {
-	// TypeTiny, TypeShort, TypeInt32 is encoded as int32
-	// TypeLong is encoded as int32 if signed, else int64.
-	// TypeLongLong is encoded as int64 if signed, else uint64,
-	// if bigintUnsignedHandlingMode set as string, encode as string.
-	case mysql.TypeTiny, mysql.TypeShort, mysql.TypeLong, mysql.TypeLonglong, mysql.TypeInt24, mysql.TypeYear:
-		switch a := value.(type) {
-		case int32:
-			buf = binary.LittleEndian.AppendUint64(buf, uint64(a))
-		case uint32:
-			buf = binary.LittleEndian.AppendUint64(buf, uint64(a))
-		case int64:
-			buf = binary.LittleEndian.AppendUint64(buf, uint64(a))
-		case uint64:
-			buf = binary.LittleEndian.AppendUint64(buf, a)
-		case string:
-			v, err := strconv.ParseUint(a, 10, 64)
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			buf = binary.LittleEndian.AppendUint64(buf, v)
-		default:
-			log.Panic("unknown golang type for the integral value",
-				zap.Any("value", value), zap.Any("mysqlType", mysqlType))
-		}
-	// TypeFloat encoded as float32, TypeDouble encoded as float64
-	case mysql.TypeFloat, mysql.TypeDouble:
-		var v float64
-		switch a := value.(type) {
-		case float32:
-			v = float64(a)
-		case float64:
-			v = a
-		}
-		if math.IsInf(v, 0) || math.IsNaN(v) {
-			v = 0
-		}
-		buf = binary.LittleEndian.AppendUint64(buf, math.Float64bits(v))
-	// TypeEnum, TypeSet encoded as string
-	// but convert to int by the getColumnValue function
-	case mysql.TypeEnum, mysql.TypeSet:
-		buf = binary.LittleEndian.AppendUint64(buf, value.(uint64))
-	// TypeBit encoded as bytes
-	case mysql.TypeBit:
-		// bit is store as bytes, convert to uint64.
-		v, err := common.BinaryLiteralToInt(value.([]byte))
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		buf = binary.LittleEndian.AppendUint64(buf, v)
-	// encoded as bytes if binary flag set to true, else string
-	case mysql.TypeVarchar, mysql.TypeVarString, mysql.TypeString, mysql.TypeTinyBlob, mysql.TypeMediumBlob, mysql.TypeLongBlob, mysql.TypeBlob:
-		switch a := value.(type) {
-		case string:
-			buf = appendLengthValue(buf, []byte(a))
-		case []byte:
-			buf = appendLengthValue(buf, a)
-		default:
-			log.Panic("unknown golang type for the string value",
-				zap.Any("value", value), zap.Any("mysqlType", mysqlType))
-		}
-	// all encoded as string
-	case mysql.TypeTimestamp, mysql.TypeDatetime, mysql.TypeDate, mysql.TypeDuration, mysql.TypeNewDate:
-		v := value.(string)
-		buf = appendLengthValue(buf, []byte(v))
-	// encoded as string if decimalHandlingMode set to string, it's required to enable checksum.
-	case mysql.TypeNewDecimal:
-		buf = appendLengthValue(buf, []byte(value.(string)))
-	// encoded as string
-	case mysql.TypeJSON:
-		buf = appendLengthValue(buf, []byte(value.(string)))
-	// this should not happen, does not take into the checksum calculation.
-	case mysql.TypeNull, mysql.TypeGeometry:
-		// do nothing
-	default:
-		return buf, errors.New("invalid type for the checksum calculation")
-	}
-	return buf, nil
-}
-
-func appendLengthValue(buf []byte, val []byte) []byte {
-	buf = binary.LittleEndian.AppendUint32(buf, uint32(len(val)))
-	buf = append(buf, val...)
-	return buf
 }
